+++ src/api_client.cpp
@@ -10,11 +10,15 @@
 
 const std::string ApiClient::BASE_URL("http://api.loda-lang.org/miner/v1/");
 
-ApiClient::ApiClient() : session_id(0), start(0), count(0) {}
+ApiClient::ApiClient()
+    : client_id(Random::get().gen() % 100000),
+      session_id(0),
+      start(0),
+      count(0) {}
 
 void ApiClient::postProgram(const Program& program) {
-  // TODO: store in tmp folder
-  const std::string tmp = "tmp_program.asm";
+  const std::string tmp =
+      getTmpDir() + "post_program_" + std::to_string(client_id) + ".asm";
   std::ofstream out(tmp);
   ProgramUtil::print(program, out);
   out.close();
@@ -86,8 +90,10 @@ void ApiClient::updateSession() {
                          std::to_string(new_count),
                      true);
   }
-  // Log::get().debug("old session:" + std::to_string(session_id) + ", old start:" + std::to_string(start) + ", old count:" +std::to_string(count) );
-  // Log::get().debug("new session:" + std::to_string(new_session_id) + ", new count:" +std::to_string(new_count) );
+  // Log::get().debug("old session:" + std::to_string(session_id) + ", old
+  // start:" + std::to_string(start) + ", old count:" +std::to_string(count) );
+  // Log::get().debug("new session:" + std::to_string(new_session_id) + ", new
+  // count:" +std::to_string(new_count) );
   start = (new_session_id == session_id) ? count : 0;
   count = new_count;
   session_id = new_session_id;
@@ -96,7 +102,9 @@ void ApiClient::updateSession() {
   for (int64_t i = 0; i < delta_count; i++) {
     queue[i] = start + i;
   }
-  // Log::get().debug("updated session:" + std::to_string(session_id) + ", updated start:" + std::to_string(start) + ", updated count:" +std::to_string(count) + " queue: " + std::to_string(queue.size()));
+  // Log::get().debug("updated session:" + std::to_string(session_id) + ",
+  // updated start:" + std::to_string(start) + ", updated count:"
+  // +std::to_string(count) + " queue: " + std::to_string(queue.size()));
   std::shuffle(queue.begin(), queue.end(), Random::get().gen);
 }